test(ai): share Vitest config
Move common node and edge Vitest settings into one helper while preserving the existing package scripts.

Fixes #15405
